LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-typed-om-expected.txt


FAIL anchor() computes to pixels assert_equals: expected 35 but got 0
FAIL anchor-size() computes to pixels assert_true: expected true got false
